Sound Detail Screen Added and functionality of playing the sounds with different volumes also added
