--- app/src/main/java/com/durranitech/presentation/ui/categoryDetailScreen.kt
@@ -1,8 +1,0 @@
-package com.durranitech.presentation.ui
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun categoryDetailScreen(){
-
-}
