+++ app/src/main/java/com/durranitech/sleepandmeditation/MainActivity.kt
@@ -1,20 +1,20 @@
 package com.durranitech.sleepandmeditation
 
+import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
-import com.durranitech.presentation.ui.BottomNavigationItem
-import com.durranitech.presentation.ui.DecisionScreen
-import com.durranitech.presentation.ui.theme.SleepAndMeditationTheme
+import com.durranitech.sleepandmeditation.presentation.ui.BottomNavigationItem
+import com.durranitech.sleepandmeditation.presentation.ui.DecisionScreen
+import com.durranitech.sleepandmeditation.presentation.ui.theme.SleepAndMeditationTheme
 
-
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -25,12 +25,12 @@ class MainActivity : ComponentActivity() {
                 val navController = rememberNavController()
 
                 Scaffold(modifier = Modifier.fillMaxSize(), bottomBar = {
-                    BottomNavigationItem(navController)
+                    BottomNavigationItem(navController = navController)
                 }) { paddingvalues ->
-                    Column(modifier = Modifier
-                        .fillMaxSize()
-                        .padding(paddingvalues)) { }
-                    DecisionScreen(navController)
+                    Column(modifier = Modifier.fillMaxSize()) {
+                        DecisionScreen(navController = navController)
+                    }
+
                 }
