--- app/src/main/java/com/durranitech/sleepandmeditation/data/model/SoundItem.kt
@@ -1,3 +1,0 @@
-package com.durranitech.sleepandmeditation.data.model
-
-data class SoundItem(val img: Int, val title: String, val imgId: Int? = null)
