+++ app/src/main/java/com/durranitech/sleepandmeditation/model/SoundItem.kt
@@ -1,0 +1,9 @@
+package com.durranitech.sleepandmeditation.model
+
+data class SoundItem(
+    val id: Int,
+    val iconResId: Int,
+    val title: String,
+    val soundResId: Int,
+    var soundVolume: Float = 1.0f
+)
