+++ app/src/main/java/com/durranitech/sleepandmeditation/repository/SoundAndIconRepository.kt
@@ -1,0 +1,6 @@
+package com.durranitech.sleepandmeditation.repository
+
+import androidx.lifecycle.ViewModel
+
+class SoundAndIconRepository: ViewModel() {
+}
